clean up versions

+++ src/compiler/rust/canister_methods/src/upload_file_chunk/versions.rs
@@ -2,42 +2,49 @@ use quote::quote;
 
 pub fn get_check_if_latest_version_src() -> proc_macro2::TokenStream {
     quote! {
-        pub fn check_if_latest_version(dest_path: &str, current_timestamp: u64) -> bool {
-            let last_recorded_timestamp = FILE_INFO.with(|uploaded_file_timestamps_map| {
-                match uploaded_file_timestamps_map.borrow().get(dest_path) {
-                    Some((timestamp, _, _)) => timestamp.clone(),
-                    None => 0,
-                }
-            });
+        pub fn check_if_latest_version(dest_path: &str, current_timestamp: Timestamp) -> bool {
+            let last_recorded_timestamp = get_timestamp(dest_path);
+
             if current_timestamp < last_recorded_timestamp {
                 // The request is from an earlier upload attempt. Disregard
                 return false;
             }
+
             if current_timestamp > last_recorded_timestamp {
                 // The request is from a newer upload attempt. Clean up the previous attempt.
-                if let Err(err) = reset_for_new_upload(dest_path, current_timestamp) {
-                    // TODO error handling?
-                    // These files may clutter the file system and take up space but otherwise are harmless.
-                    // Unless they don't get overwritten properly and the chunk-writer tries to use them to create the next version of this file.
-                    ic_cdk::println!(
-                        "WARNING: Failed to delete file {} uploaded at {}. {}",
-                        dest_path,
-                        current_timestamp,
-                        err
-                    );
-                }
+                reset_for_new_upload(dest_path, current_timestamp).unwrap();
             }
+
             true
         }
 
-        fn reset_for_new_upload(path: &str, timestamp: u64) -> std::io::Result<()> {
+        fn reset_for_new_upload(path: &str, timestamp: Timestamp) -> std::io::Result<()> {
+            initialize_file_info(path, timestamp);
+            delete_if_exists(path)?;
+            Ok(())
+        }
+
+        fn initialize_file_info(path: &str, timestamp: Timestamp) {
             FILE_INFO.with(|file_info| {
                 let mut file_info_mut = file_info.borrow_mut();
 
                 file_info_mut.insert(path.to_string(), (timestamp, 0, vec![]));
             });
-            // TODO check if the file exists
-            std::fs::remove_file(path)?;
+        }
+
+        fn get_timestamp(path: &str) -> Timestamp {
+            FILE_INFO.with(|uploaded_file_timestamps_map| {
+                match uploaded_file_timestamps_map.borrow().get(path) {
+                    Some((timestamp, _, _)) => timestamp.clone(),
+                    None => 0,
+                }
+            })
+        }
+
+        fn delete_if_exists(path: &str) -> std::io::Result<()> {
+            if std::fs::metadata(path).is_ok() {
+                std::fs::remove_file(path)?;
+            }
             Ok(())
         }
